BT, Loop, and GPS changes
BT - It works now, I still have to test outside
Loop - added time for BT sending at the beginning. Loop is now 120 ms
GPS - changed leader and follow time to unsigned longs

+++ libraries/main/SendGPS.h
@@ -16,8 +16,8 @@ public:
 
   float lat = 0;
   float lon = 0;
-  float leaderTime;
-  float followTime;
+  unsigned long leaderTime;
+  unsigned long  followTime;
 
   void updateState(float lat, float lon, unsigned long leaderTime, unsigned long followTime);
